feat: 알림 기업 추가 API 개발
NotiHistoryEntity 생성
UserService로 userId 가져옴
Notification - service, repository 추가

Close #24

+++ src/main/java/org/bob/siungongsi/controller/dto/NotificationRequest.java
@@ -1,6 +1,5 @@
 package org.bob.siungongsi.controller.dto;
 
 public class NotificationRequest {
-
-  public record NotificationCompanyRequest(int companyId) {}
+  public record NotificationCompanyRequest(Long companyId) {}
 }
